refactoring of crypto
